RFPL-2088: Update cargo plugin to work with JBoss EAP6
Update cargo plugin version, some configuration options accrdingly,
and update configuration for JBoss AS to 6.0.0.Final, as it is last stable
version.
Main change is profile for JBoss EAP6 (based on JBoss AS 7.1.x), since
cargo plugin 1.2.0 support it. Verified that at least with EAP6 it is possible run whole surefire suite.